Improve gradle plugins again
