+++ kmp/kmp-common-test/build.gradle.kts
@@ -1,7 +1,7 @@
-import com.gchristov.thecodinglove.gradleplugins.Deps
+import com.gchristov.thecodinglove.kmpgradleplugins.Deps
 
 plugins {
-    id("javascript-node-target-plugin")
+    id("kmp-platform-plugin")
 }
 
 kotlin {
