--- kmp/kmp-gradle-plugins/src/main/kotlin/com/gchristov/thecodinglove/gradleplugins/js/JavascriptTargetPlugin.kt
@@ -1,33 +1,0 @@
-package com.gchristov.thecodinglove.gradleplugins.js
-
-import com.gchristov.thecodinglove.gradleplugins.KmpPlatformPlugin
-import org.gradle.api.Project
-import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
-
-@Suppress("unused")
-open class JavascriptBrowserTargetPlugin : KmpPlatformPlugin() {
-    override fun apply(target: Project) {
-        super.apply(target)
-        target.extensions.configure(KotlinMultiplatformExtension::class.java) {
-            js(IR) {
-                browser {
-                    commonWebpackConfig {
-                        cssSupport.enabled = true
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Suppress("unused")
-open class JavascriptNodeTargetPlugin : KmpPlatformPlugin() {
-    override fun apply(target: Project) {
-        super.apply(target)
-        target.extensions.configure(KotlinMultiplatformExtension::class.java) {
-            js(IR) {
-                nodejs()
-            }
-        }
-    }
-}
