+++ kmp/kmp-gradle-plugins/src/main/kotlin/com/gchristov/thecodinglove/kmpgradleplugins/BuildConfigPlugin.kt
@@ -1,4 +1,4 @@
-package com.gchristov.thecodinglove.gradleplugins
+package com.gchristov.thecodinglove.kmpgradleplugins
 
 import org.gradle.api.Plugin
 import org.gradle.api.Project
