+++ kmp/kmp-gradle-plugins/src/main/kotlin/com/gchristov/thecodinglove/kmpgradleplugins/KmpPlatformPlugin.kt
@@ -1,0 +1,48 @@
+package com.gchristov.thecodinglove.kmpgradleplugins
+
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
+
+@Suppress("unused")
+class JavascriptBrowserExecutablePlugin : Plugin<Project> {
+    override fun apply(target: Project) {
+        target.plugins.apply("org.jetbrains.kotlin.multiplatform")
+        target.extensions.configure(KotlinMultiplatformExtension::class.java) {
+            js(IR) {
+                binaries.executable()
+                browser {
+                    commonWebpackConfig {
+                        cssSupport.enabled = true
+                    }
+                }
+            }
+        }
+    }
+}
+
+@Suppress("unused")
+class JavascriptNodeExecutablePlugin : Plugin<Project> {
+    override fun apply(target: Project) {
+        target.plugins.apply("org.jetbrains.kotlin.multiplatform")
+        target.plugins.apply("dev.petuska.npm.publish")
+        target.extensions.configure(KotlinMultiplatformExtension::class.java) {
+            js(IR) {
+                binaries.library()
+                nodejs()
+            }
+        }
+    }
+}
+
+@Suppress("unused")
+abstract class KmpPlatformPlugin : Plugin<Project> {
+    override fun apply(target: Project) {
+        target.plugins.apply("org.jetbrains.kotlin.multiplatform")
+        target.extensions.configure(KotlinMultiplatformExtension::class.java) {
+            js(IR) {
+                nodejs()
+            }
+        }
+    }
+}
