+++ kmp/kmp-module-b/build.gradle.kts
@@ -1,3 +1,3 @@
 plugins {
-    id("javascript-node-module-plugin")
+    id("kmp-module-plugin")
 }
